% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractDist.R
\name{extractDist}
\alias{extractDist}
\title{Subset a symmetric distance matrix of class "dist"}
\usage{
extractDist(dist, idxRows = NULL, idxCols = NULL)
}
\arguments{
\item{dist}{A "dist" object, which can be obtained by the "dist" function.}

\item{idxRows}{An interger vector indicates the indexes of units in the first subset. It cannot be smaller than 1
or larger than the dataset size. By default, idxRows = NULL.}

\item{idxCols}{An interger vector indicates the indexes of units in the second subset. It cannot be smaller than 1
or larger than the dataset size. By default, idxCols = NULL.}
}
\value{
A numeric matrix storing pair-wise distances between the units in each subset.
}
\description{
extractDist() allows us to efficiently subset a distance matrix of class "dist" without converting
the "dist" object to a numeric matrix.
}
\details{
We are interested in extracting pair-wises distances between units in two subsets (indicated by their
indexes) from a "dist" object. A simple way to do that is to convert the "dist" object to a numeric matrix using
as.dist() and extract the relevant values using the squared brackets operator for matrix subsetting. However,
it is very slow. extractDist() efficiently extract the pair-wise distances from the "dist" object without the need of conversion.

When either idxRows or idxCols are NULL, extractDist() extracts column vectors of the "dist" matrix given by
the indexes of the not-null vector. Since the distance matrix is symmetric, it does not matter if we extract
the row vectors or the column vectors. However, our implementation is more efficient for extracting column vectors
from a "dist" object. If idxRows and idxCols are both NULL, the "dist" object is converted to a full numeric
distance matrix.
}
\examples{

x = rnorm(100)
dx = dist(x)
extractDist(dx, idxRows = 1:10, idxCols = 25:50)

}
\author{
Minh Long Nguyen \email{edelweiss611428@gmail.com}
}
