% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subDist.R
\name{subDist}
\alias{subDist}
\title{Extract a sub-distance matrix from a "dist" object}
\usage{
subDist(dist, idx, diag = F, upper = F)
}
\arguments{
\item{dist}{A "dist" object, which can be obtained by the "dist" function.}

\item{idx}{An integer vector indicates the indexes of the extracted units (idx can't be smaller than 1 or larger
than the dataset size N.}

\item{diag}{A boolean value indicates whether or not attr(dist, "Diag") = TRUE (by default, FALSE).}

\item{upper}{A boolean value indicates whether or not attr(dist, "Upper") = TRUE (by default, FALSE).}
}
\value{
a sub-distance matrix of class "dist"
}
\description{
subDist() allows us to efficiently extract from a "dist" object a sub-distance matrix of class "dist"
of a subset of the original dataset.
}
\details{
We are interested in extracting from a "dist" object a sub-distance matrix of class "dist" of a subset
of the original dataset. A simple way to do that is to convert the "dist" object to a numeric matrix using as.matrix(),
extract the relevant values, and convert the matrix back to a "dist" object using as.dist(). However, it is slow.
subDist() allows us to do that efficiently without the need of conversion.
}
\examples{

x = rnorm(1000)
dx = dist(x)
subDist(dx, 1:10)

}
\author{
Minh Long Nguyen \email{edelweiss611428@gmail.com}
}
